pyproject.toml
[project]
name = "trendingpost"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.12.13",
    "beautifulsoup4>=4.13.4",
    "duckduckgo-search>=8.0.5",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.27",
    "langchain-xai>=0.2.4",
    "langgraph>=0.5.1",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.9.0",
    "transformers>=4.53.0",
]
